Allow variable delay between pings
This is meant to support small and large numbers of points in a rendered
dataset. For smaller values, discrete pings will be heard. As sizes
grow, it will sound continuous. Note that the overall size of the
rendering will only begin increasing after 1000 points, and this may
need to be revised down (with ping delay markers adjusted) depending on
encountered examples.

Resolves #842.